too many things

+++ main.py
@@ -9,7 +9,7 @@ import pandas as pd
 
 # import tqdm
 
-from utils.pathutils import get_snowflake_id_from_path, get_image_paths, get_image_filename, get_filtered_image_paths, __deprecated___get_image_paths_filtered
+from utils.pathutils import get_snowflake_id_from_path, get_image_paths, get_image_filename, get_filtered_image_paths, __deprecated___get_image_paths_filtered, get_image_folder, get_str_image_id_from_path
 from utils.plotutils import plot_ellipse_overlay
 from utils.consolecolors import bcolors
 from utils.utils import info, warn, err, header
@@ -20,7 +20,9 @@ from metrics import analyse_image
 # dir_list = ["10-7_15-42-5/", "10-7_15-45-41/", "10-7_15-51-6/", "10-28_15-27-6"]
 # path = base_path + "/" + dir_list[-1]
 
-save_path = "/mnt/e/snowflake_analysis_plots/run-2"
+base_path = "/mnt/f/snowflake_analysis_plots"
+run_number = time.strftime("%Y-%m-%d_%a_%H-%M-%S", time.localtime())
+save_path = base_path + "/" + run_number
 if not os.path.exists(save_path):
     os.makedirs(save_path, exist_ok=True)
 
@@ -30,16 +32,35 @@ if __name__=="__main__":
     csv_filename = "snowflake_image_metrics.csv"
     csv_filepath = os.path.join(save_path, csv_filename)
     if os.path.exists(csv_filepath):
-        warn(f"CSV file {csv_filepath} already exists and will be overwritten.")
+        warn(f"CSV file {csv_filepath} already exists and will be overwritten. Are you sure? [y/N]")
+        user_input = input().strip().lower()
+        if user_input != 'y':
+            info("Exiting program.")
+            exit(0)        
         os.remove(csv_filepath)
         
     df = pd.DataFrame(columns=['min_intensity', 'max_intensity', 'mean_intensity', 'std_intensity', 'variance_intensity', 'entropy', 'snr', 'mean_gradient_magnitude', 'median_gradient_magnitude', 'std_gradient_magnitude', 'mean_laplace', 'median_laplace', 'std_laplace', 'mean_angle', 'median_angle', 'std_angle', 'sharp_edge_count'] + ['image_name'])
     
     # images = get_image_paths("/mnt/e/pictures_Vikram/10-29_11-52-33/")
     # images = get_filtered_image_paths()
-    # images = __deprecated___get_image_paths_filtered(image_dir="/mnt/e/pictures_Test_old/", start_from="10-22_12-27-20")
-    images = get_image_paths("../images/nice_flakes")
+    # images = __deprecated___get_image_paths_filtered(image_dir="/mnt/e/pictures_Test/", start_from="11-12_15-50-21") # test_old, 10-22_12-27-20
+    # images = get_image_paths("../images/nice_flakes")
     # images = ["../../images/nice_flakes/Snowflake_20.bmp"] #
+    images = get_image_paths("/mnt/f/davos_1/11-17_1-57-46/") # 11-17_2-29-24
+    # images = ["/mnt/f/davos_1/11-17_2-29-24/Snowflake_284.bmp"]
+    
+    '''
+    NOTE: When imaging with a bright background, ensure to invert the image colors before processing. The snowflake should appear brighter than the background for accurate analysis.
+    
+    # Example of inverting an image using OpenCV
+    image = cv2.imread('path_to_image', cv2.IMREAD_GRAYSCALE)   
+    inverted_image = cv2.bitwise_not(image)
+    '''
+    
+    # sort images by snowflake id extracted from filename
+    images = [img for _, img in sorted((get_snowflake_id_from_path(img), img) for img in images)]
+    
+    
     try:
         from tqdm import tqdm
         images_iterable = tqdm(images)
@@ -52,20 +73,22 @@ if __name__=="__main__":
             name = get_image_filename(img_path)
             if name.startswith("processed_"):
                 continue
-            header(f"Processing image: {img_path}, id: {get_snowflake_id_from_path(img_path)}")
+            header(f"Processing image: {img_path}, id: {get_snowflake_id_from_path(img_path)} from folder: {get_image_folder(img_path)} with str id: {get_str_image_id_from_path(img_path)}")
             
             res: Optional[MatLike] = cv2.imread(img_path, cv2.IMREAD_GRAYSCALE)
             if res is None:
                 err(f"Failed to read image: {img_path}")
                 continue
             
-            # res = cv2.createCLAHE(clipLimit=2.0, tileGridSize=(8,8)).apply(res)
-            # res = gamma(res, gamma=0.8)
-            # cv2.imshow("CLAHE Result", res)
-            # cv2.waitKey(10000)
+            frame: MatLike = res.copy()
+            cv2.createCLAHE(clipLimit=2.0, tileGridSize=(8,8)).apply(src=frame)
+            frame = gamma(frame, gamma=0.8)
+            cv2.imshow("CLAHE Result", frame)
+            cv2.waitKey(1)
+            
             processed_image, elapsed_processor  = preprocess_image(res)
             stats                               = image_stats(res) # returns an np array
-            metrics, elapsed_analyser           = analyse_image(res, visual=True, save=True, save_path=save_path)
+            metrics, elapsed_analyser           = analyse_image(image=res, plot=True, display_plot=False, save=True, save_path=save_path, folder_desc=f"{get_str_image_id_from_path(img_path)}")
             print(f"Metrics for {os.path.basename(img_path)}: {metrics}")
             
             df.loc[len(df)] = np.concatenate((stats, [name]))
